rename: 문자열 상수 디렉토리 구조 변경
- ConsoleMessages
  - Console에 의해 입출력되는 문자열 상수 정의
- GameResultMessageGenerator
  - 게임 결과에 따른 문자열 생성기
- GameResults
  - 게임 결과 문자열 상수 정의
