+++ src/main/java/baseball/util/ConsoleMessages.java
@@ -1,0 +1,8 @@
+package baseball.util;
+
+public class ConsoleMessages {
+    public static final String GAME_START = "숫자 야구 게임을 시작합니다.";
+    public static final String INPUT_NUMBER = "숫자를 입력해주세요 : ";
+    public static final String GAME_OVER = "3개의 숫자를 모두 맞히셨습니다! 게임 종료";
+    public static final String RESTART_OR_TERMINATE = "게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.";
+}
