+++ src/main/java/baseball/util/GameResultMessageGenerator.java
@@ -1,13 +1,8 @@
-package baseball.game;
+package baseball.util;
 
-public class BaseBallGameMessage {
-    public static final String GAME_START = "숫자 야구 게임을 시작합니다.";
-    public static final String INPUT_NUMBER = "숫자를 입력해주세요 : ";
-    public static final String GAME_OVER = "3개의 숫자를 모두 맞히셨습니다! 게임 종료";
-    public static final String RESTART_OR_TERMINATE = "게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.";
-    private static final String BALL = "볼 ";
-    private static final String STRIKE = "스트라이크";
-    private static final String NOTHING = "낫싱";
+import static baseball.util.GameResults.*;
+
+public class GameResultMessageGenerator {
 
 
     public static String makeResultString(int strike, int ball) {
