+++ src/main/java/baseball/util/GameResults.java
@@ -1,0 +1,17 @@
+package baseball.util;
+
+public class GameResults {
+
+    public static final String BALL = "볼 ";
+    public static final String STRIKE = "스트라이크";
+    public static final String NOTHING = "낫싱";
+
+
+    public static final String ONE_BALL = "1볼 ";
+    public static final String TWO_BALL = "2볼 ";
+    public static final String THREE_BALL = "3볼 ";
+    public static final String ONE_STRIKE = "1스트라이크";
+    public static final String TWO_STRIKE = "2스트라이크";
+    public static final String THREE_STRIKE = "3스트라이크";
+    public static final String GAME_OVER = "게임 종료";
+}
